feat: add opt-in usage metrics
This adds posthog as a service to collect some usage data around gadgets. The feature
is opt-in (nothing is collected by default) and currently only captures the names of
gadgets so we can get some numbers around which gadgets are used more frequently.
This might be extended or removed completely in the future.

Signed-off-by: Michael Friese <mfriese@microsoft.com>